decomp: match batch (+6 functions, 136/633 total)

+++ src/matched/sub_08067634.c
@@ -1,7 +1,7 @@
 #include "global.h"
 
 // @ 0x08067634
-u16 sub_08067634(u16 a, u32 b)
+u16 sub_08067634(u32 a, u32 b)
 {
     return sub_08067648(a, b, 1);
 }
